requirements.txt
torch>=2.1.0
pyyaml>=6.0
tqdm>=4.65.0
safetensors>=0.4.0
tokenizers>=0.19.0
transformers>=4.40.0
datasets>=2.18.0
packaging
requests>=2.31.0
pre-commit>=3.5.0
openai>=1.0.0        # optional: only needed for --provider openai
python-dotenv>=1.0.0 # optional: loads .env file for OPENAI_API_KEY
